sm-bootstrap/requirements.txt
aws-cdk.core==1.114.0
aws-cdk.aws-sagemaker==1.114.0
